feat!: consolidate Babylon Lite exports under root entry
Remove package subpath exports, expose supported APIs from the root barrel, migrate consumers, and enforce the single-entry contract in build tests and guidance.

BREAKING CHANGE: Package subpath exports were removed. Import all public APIs from @babylonjs/lite instead.

Copilot-Session: 9c1a46a8-43d6-4e91-ad55-475045ec423b
