+++ lab/lite/src/lite/scene179.ts
@@ -6,14 +6,16 @@ import {
     addToScene,
     attachFreeControl,
     createEngine,
+    createClusteredLightContainer,
+    createClusteredPointLight,
     createFreeCamera,
     createSceneContext,
+    addClusteredLightContainer,
     loadGltf,
     registerScene,
     startEngine,
     type PbrMaterialProps,
 } from "babylon-lite";
-import { addClusteredLightContainer, createClusteredLightContainer, createClusteredPointLight } from "babylon-lite/light/clustered";
 
 const MODEL_URL = "https://cdn.jsdelivr.net/gh/KhronosGroup/glTF-Sample-Models@master/2.0/Sponza/glTF/Sponza.gltf";
